drug-drug interaction/requirements.txt
groq
numpy
requests
python-dotenv
